Added check for file linking

+++ start.sh
@@ -8,13 +8,16 @@ echo 'WIFI_SSID = "'"$WIFI_SSID"'" ' >> config.py
 echo 'WIFI_PASS = "'"$WIFI_PASS"'" ' >> config.py
 
 # Remove config files if they exist
-rm Pycom/1.0.0/flash/config.py
-rm Pycom/1.0.1/flash/config.py
+if [ Pycom/1.0.0/flash/config.py -o Pycom/1.0.1/flash/config.py ]; then
+    rm Pycom/1.0.0/flash/config.py
+    rm Pycom/1.0.1/flash/config.py
+fi
 
 # Link configuration script to SW stores
 ln config.py Pycom/1.0.0/flash/config.py
 ln config.py Pycom/1.0.1/flash/config.py
 
+echo "Starting OTA Server"
 python Pycom/OTA_server.py
 
 function cleanup {
